Revert Async changes to SNIPacket to fix deadlock issues
Fixes #262 by reverting dotnet/corefx#34184 until PR #335 is ready with complete fix.
